Fix HoverPluginImpl
