+++ src/index.ts
@@ -1,2 +1,3 @@
 export { Converter } from "./Converter";
 export { HoverableE } from "./HoverableE";
+export { HoverPluginImpl as HoverPlugin } from "./HoverPluginImpl";
